Move styling logic to renderer
